+++ src/sections/pandora-browser.ts
@@ -71,7 +71,7 @@ export class PandoraBrowser extends FavBrowserBase {
           ${this.alertError ? html`<ha-alert alert-type="error" dismissable @alert-dismissed-clicked=${this.alertErrorClear}>${this.alertError}</ha-alert>` : ""}
           ${this.alertInfo ? html`<ha-alert alert-type="info" dismissable @alert-dismissed-clicked=${this.alertInfoClear}>${this.alertInfo}</ha-alert>` : ""}
           ${(() => {
-            if (this.config.pandoraBrowserItemsPerRow === 1) {
+            if ((this.config.pandoraBrowserItemsPerRow || 1) === 1) {
               return (
                 html`<stpc-media-browser-list
                       class="media-browser-list"